Compile tinygettext as a static library project with the game instead of building it seperately.
Tinygettext uses stl types in the interface. This can cause crashes if tinygettext is compiled with different compilers than the game or sometimes even if the compiler settings differ a bit.
Building it with the game should ensure that the compiler and compiler settings are the same.
I've tested this version on Linux, but it probably needs some tweaks for Windows. I'm going to test it there too soon.