Fixed the FormatMillisecondsIntoDateString engine function.
The engine was expecting an integer when it actually should expect a UDate, which is a ICU typedef for a double.
Once the engine received the milliseconds as an UDate, the L10n.cpp code was simplified and fixed.
I also deleted a call to Engine.FormatMillisecondsIntoDateString() in JavaScript that was not being used at all (the return value was ignored). Probably a simple leftover from a codign session.
Modified some code to use Engine.FormatMillisecondsIntoDateString() instead of concatenating hours and minutes.
Added *1000 to fix a case where the timestamp was given in seconds instead of milliseconds.